Updated some definitions of the Chip-8 object and added text sprite data
Added a new config folder containing files that define basic information for the Chip-8. In this revision I added a textSprites.bin file that contains all the text sprite data for the dafualt text sprites. This will be read into memory and a pointer passed to uint8_t *textSprites[][] during chip-8 initialization